+++ lib/gui/app.js
@@ -40,7 +40,7 @@ const settings = require('./models/settings');
 const windowProgress = require('./os/window-progress');
 const analytics = require('./modules/analytics');
 const updateNotifier = require('./components/update-notifier');
-const availableDrives = require('./models/drives');
+const availableDrives = require('./models/available-drives');
 
 const Store = require('./models/store');
