+++ lib/gui/components/drive-selector/controllers/drive-selector.js
@@ -21,7 +21,7 @@ const _ = require('lodash');
 const messages = require('../../../../shared/messages');
 const constraints = require('../../../../shared/drive-constraints');
 const analytics = require('../../../modules/analytics');
-const availableDrives = require('../../../models/drives');
+const availableDrives = require('../../../models/available-drives');
 
 module.exports = function(
   $q,
