+++ lib/gui/models/available-drives.js
