+++ lib/gui/models/selection-state.js
@@ -23,7 +23,7 @@
 const _ = require('lodash');
 const angular = require('angular');
 const Store = require('./store');
-const availableDrives = require('./drives');
+const availableDrives = require('./available-drives');
 const MODULE_NAME = 'Etcher.Models.SelectionState';
 const SelectionStateModel = angular.module(MODULE_NAME, []);
