+++ lib/gui/pages/main/controllers/main.js
@@ -19,7 +19,7 @@
 const settings = require('../../../models/settings');
 const flashState = require('../../../models/flash-state');
 const analytics = require('../../../modules/analytics');
-const availableDrives = require('../../../models/drives');
+const availableDrives = require('../../../models/available-drives');
 
 module.exports = function(
   SelectionStateModel,
