+++ tests/gui/models/drives.spec.js
@@ -3,7 +3,7 @@
 const m = require('mochainon');
 const path = require('path');
 const angular = require('angular');
-const availableDrives = require('../../../lib/gui/models/drives');
+const availableDrives = require('../../../lib/gui/models/available-drives');
 require('angular-mocks');
 
 describe('Browser: availableDrives', function() {
